feat: cleaned up docs and API functionality
