+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "icefabric"
-version = "2025.5"
+version = "2025.6.0"
 description = "An Apache Iceberg implementation of the Hydrofabric"
 readme = "README.md"
 requires-python = ">=3.11,<3.13"
