+++ src/icefabric_api/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "icefabric-api"
-version = "2025.5.0"
+name = "icefabric_api"
+version = "2025.6.0"
 description = "A FastAPI API for querying apache iceberg tables"
 readme = "README.md"
 requires-python = ">=3.11, <3.13"
@@ -9,3 +9,7 @@ dependencies = [
     "pydantic==2.11.5",
     "pydantic-settings==2.9.1",
 ]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
