+++ src/icefabric_api/src/icefabric_api/__init__.py
@@ -1,0 +1,6 @@
+from icefabric_api.create_ipes import get_hydrofabric_attributes, module_ipe
+
+__all__ = [
+    "get_hydrofabric_attributes",
+    "module_ipe",
+]
