--- src/icefabric_api/src/requirements.txt
@@ -1,2 +1,0 @@
-geopandas==1.0.1
-pandas==2.2.3
