+++ src/icefabric_manage/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "icefabric_manage"
-version = "2025.5"
+version = "2025.6.0"
 description = "All functionality for creating, updating, and deleting information from the icefabric catalog"
 authors = [
   { name = "Tadd Bindas", email = "tadd.bindas@ertcorp.com" },
