+++ src/icefabric_tools/src/icefabric_tools/__init__.py
@@ -2,4 +2,11 @@ from icefabric_tools.iceberg.hydrofabric import IdType, subset
 from icefabric_tools.iceberg.rnr import get_rnr_segment
 from icefabric_tools.iceberg.utils import find_origin, table_to_geopandas, to_geopandas
 
-__all__ = ["IdType", "subset", "get_rnr_segment", "find_origin", "to_geopandas", "table_to_geopandas"]
+__all__ = [
+    "IdType",
+    "subset",
+    "get_rnr_segment",
+    "find_origin",
+    "to_geopandas",
+    "table_to_geopandas",
+]
